Backend/requirements.txt
certifi==2021.5.30
chardet==4.0.0
Click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.26.0
stripe==6.0.0
toml==0.10.2
Werkzeug==2.0.1
tiktoken
rq
langchain
openai
flask_cors
bs4
redis
sendgrid
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
flask-migrate
